Split monolithic index.html into separate files under public/
Move CSS, ADB module, and app JS out of the 2080-line single HTML file
into public/style.css, public/adb.js, and public/app.js. Update Flask
to serve static files from public/ directory.
