+++ public/app.js
@@ -1,1219 +1,3 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=no">
-  <meta name="theme-color" content="#0a0a0a">
-  <title>GPlay Downloader</title>
-  <link rel="preconnect" href="https://fonts.googleapis.com">
-  <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
-  <link href="https://fonts.googleapis.com/css2?family=JetBrains+Mono:wght@400;500;600;700&family=Plus+Jakarta+Sans:wght@400;500;600;700&display=swap" rel="stylesheet">
-  <script src="https://cdnjs.cloudflare.com/ajax/libs/jszip/3.10.1/jszip.min.js" integrity="sha384-+mbV2IY1Zk/X1p/nWllGySJSUN8uMs+gUAN10Or95UBH0fpj6GfKgPmgC5EXieXG" crossorigin="anonymous"></script>
-  <style>
-    :root {
-      --bg-primary: #0a0a0a;
-      --bg-card: #111113;
-      --bg-card-hover: #161618;
-      --bg-input: #0e0e10;
-      --bg-elevated: #1a1a1e;
-      --border: #222228;
-      --border-focus: #3d3d44;
-      --text-primary: #e4e4e7;
-      --text-secondary: #71717a;
-      --text-tertiary: #6e6e77;
-      --accent: #2563eb;
-      --accent-dim: #1d4ed8;
-      --accent-glow: rgba(37, 99, 235, 0.15);
-      --accent-glow-strong: rgba(37, 99, 235, 0.3);
-      --success: #4ade80;
-      --success-bg: rgba(74, 222, 128, 0.08);
-      --error: #f87171;
-      --error-bg: rgba(248, 113, 113, 0.08);
-      --info: #60a5fa;
-      --info-bg: rgba(96, 165, 250, 0.08);
-      --radius: 10px;
-      --radius-lg: 14px;
-      --font-mono: 'JetBrains Mono', 'SF Mono', 'Fira Code', monospace;
-      --font-sans: 'Plus Jakarta Sans', -apple-system, sans-serif;
-      --transition: 150ms cubic-bezier(0.4, 0, 0.2, 1);
-    }
-
-    * { box-sizing: border-box; margin: 0; padding: 0; }
-
-    body {
-      font-family: var(--font-sans);
-      background: var(--bg-primary);
-      color: var(--text-primary);
-      min-height: 100vh;
-      -webkit-font-smoothing: antialiased;
-      -moz-osx-font-smoothing: grayscale;
-    }
-
-    /* Subtle noise texture */
-    body::before {
-      content: '';
-      position: fixed;
-      inset: 0;
-      opacity: 0.025;
-      background-image: url("data:image/svg+xml,%3Csvg viewBox='0 0 256 256' xmlns='http://www.w3.org/2000/svg'%3E%3Cfilter id='n'%3E%3CfeTurbulence type='fractalNoise' baseFrequency='0.9' numOctaves='4' stitchTiles='stitch'/%3E%3C/filter%3E%3Crect width='100%25' height='100%25' filter='url(%23n)'/%3E%3C/svg%3E");
-      background-size: 200px 200px;
-      pointer-events: none;
-      z-index: 9999;
-    }
-
-    /* Top glow accent line */
-    body::after {
-      content: '';
-      position: fixed;
-      top: 0;
-      left: 0;
-      right: 0;
-      height: 1px;
-      background: linear-gradient(90deg, transparent, var(--accent) 30%, var(--accent) 70%, transparent);
-      opacity: 0.4;
-      z-index: 100;
-    }
-
-    .wrapper {
-      max-width: 620px;
-      margin: 0 auto;
-      padding: 40px 20px 60px;
-    }
-
-    /* Header */
-    .header {
-      text-align: center;
-      margin-bottom: 40px;
-    }
-
-    .header-badge {
-      display: inline-flex;
-      align-items: center;
-      gap: 6px;
-      font-family: var(--font-mono);
-      font-size: 11px;
-      font-weight: 500;
-      letter-spacing: 0.08em;
-      text-transform: uppercase;
-      color: var(--accent-dim);
-      background: var(--accent-glow);
-      border: 1px solid rgba(163, 230, 53, 0.15);
-      padding: 5px 12px;
-      border-radius: 20px;
-      margin-bottom: 16px;
-    }
-
-    .header-badge svg {
-      width: 13px;
-      height: 13px;
-    }
-
-    .header h1 {
-      font-family: var(--font-mono);
-      font-size: clamp(1.6rem, 5vw, 2rem);
-      font-weight: 700;
-      color: var(--text-primary);
-      letter-spacing: -0.03em;
-      line-height: 1.1;
-    }
-
-    .header h1 span {
-      color: var(--accent);
-    }
-
-    .header p {
-      margin-top: 8px;
-      color: var(--text-tertiary);
-      font-size: 13px;
-    }
-
-    /* Cards */
-    .card {
-      background: var(--bg-card);
-      border: 1px solid var(--border);
-      border-radius: var(--radius-lg);
-      padding: 24px;
-      margin-bottom: 16px;
-      transition: border-color var(--transition);
-    }
-
-    .card:hover {
-      border-color: var(--border-focus);
-    }
-
-    .card-label {
-      font-family: var(--font-mono);
-      font-size: 11px;
-      font-weight: 600;
-      letter-spacing: 0.1em;
-      text-transform: uppercase;
-      color: var(--text-tertiary);
-      margin-bottom: 16px;
-      display: flex;
-      align-items: center;
-      gap: 8px;
-    }
-
-    .card-label svg {
-      width: 14px;
-      height: 14px;
-      color: var(--text-tertiary);
-    }
-
-    /* Form Controls */
-    input[type="text"] {
-      width: 100%;
-      padding: 12px 14px;
-      border: 1px solid var(--border);
-      border-radius: var(--radius);
-      background: var(--bg-input);
-      color: var(--text-primary);
-      font-family: var(--font-mono);
-      font-size: 14px;
-      font-weight: 400;
-      margin-bottom: 12px;
-      transition: border-color var(--transition), box-shadow var(--transition);
-      outline: none;
-    }
-
-    input[type="text"]::placeholder {
-      color: var(--text-tertiary);
-      font-family: var(--font-mono);
-    }
-
-    input[type="text"]:focus {
-      border-color: var(--accent-dim);
-      box-shadow: 0 0 0 3px var(--accent-glow);
-    }
-
-    input:-webkit-autofill {
-      -webkit-box-shadow: 0 0 0 30px var(--bg-input) inset !important;
-      -webkit-text-fill-color: var(--text-primary) !important;
-    }
-
-    select {
-      padding: 10px 12px;
-      border-radius: var(--radius);
-      background: var(--bg-input);
-      color: var(--text-primary);
-      border: 1px solid var(--border);
-      font-family: var(--font-sans);
-      font-size: 13px;
-      font-weight: 500;
-      flex: 1;
-      min-width: 150px;
-      outline: none;
-      cursor: pointer;
-      transition: border-color var(--transition);
-      appearance: none;
-      background-image: url("data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' width='16' height='16' viewBox='0 0 24 24' fill='none' stroke='%2371717a' stroke-width='2' stroke-linecap='round' stroke-linejoin='round'%3E%3Cpolyline points='6 9 12 15 18 9'/%3E%3C/svg%3E");
-      background-repeat: no-repeat;
-      background-position: right 10px center;
-      padding-right: 32px;
-    }
-
-    select:focus {
-      border-color: var(--accent-dim);
-      box-shadow: 0 0 0 3px var(--accent-glow);
-    }
-
-    /* Checkbox */
-    .checkbox-wrap {
-      display: flex;
-      align-items: center;
-      gap: 8px;
-      cursor: pointer;
-      user-select: none;
-      white-space: nowrap;
-    }
-
-    .checkbox-wrap input[type="checkbox"] {
-      appearance: none;
-      width: 16px;
-      height: 16px;
-      border: 1.5px solid var(--border-focus);
-      border-radius: 4px;
-      background: var(--bg-input);
-      cursor: pointer;
-      position: relative;
-      transition: all var(--transition);
-      flex-shrink: 0;
-    }
-
-    .checkbox-wrap input[type="checkbox"]:checked {
-      background: var(--accent-dim);
-      border-color: var(--accent-dim);
-    }
-
-    .checkbox-wrap input[type="checkbox"]:checked::after {
-      content: '';
-      position: absolute;
-      left: 4.5px;
-      top: 1.5px;
-      width: 5px;
-      height: 8px;
-      border: solid var(--bg-primary);
-      border-width: 0 2px 2px 0;
-      transform: rotate(45deg);
-    }
-
-    .checkbox-wrap span {
-      font-size: 13px;
-      font-weight: 500;
-      color: var(--text-secondary);
-    }
-
-    .options-row {
-      display: flex;
-      gap: 12px;
-      margin-bottom: 12px;
-      flex-wrap: wrap;
-      align-items: center;
-    }
-
-    .note {
-      font-size: 11px;
-      color: var(--text-tertiary);
-      line-height: 1.5;
-      margin-bottom: 16px;
-      font-family: var(--font-mono);
-      padding: 10px 12px;
-      background: rgba(255,255,255,0.02);
-      border-radius: var(--radius);
-      border-left: 2px solid var(--border);
-    }
-
-    /* Buttons */
-    button {
-      padding: 10px 20px;
-      border: 1px solid transparent;
-      border-radius: var(--radius);
-      font-family: var(--font-sans);
-      font-weight: 600;
-      font-size: 13px;
-      cursor: pointer;
-      transition: all var(--transition);
-      outline: none;
-      display: inline-flex;
-      align-items: center;
-      justify-content: center;
-      gap: 6px;
-    }
-
-    .btn-primary {
-      background: var(--accent);
-      color: #fff;
-      border-color: var(--accent);
-    }
-
-    .btn-primary:hover {
-      background: var(--accent-dim);
-      border-color: var(--accent-dim);
-      box-shadow: 0 0 20px var(--accent-glow);
-    }
-
-    .btn-primary:active {
-      transform: scale(0.98);
-    }
-
-    .btn-secondary {
-      background: var(--bg-elevated);
-      color: var(--text-primary);
-      border-color: var(--border);
-    }
-
-    .btn-secondary:hover {
-      background: var(--bg-card-hover);
-      border-color: var(--border-focus);
-    }
-
-    .btn-ghost {
-      background: transparent;
-      color: var(--text-secondary);
-      border-color: transparent;
-      padding: 8px 14px;
-    }
-
-    .btn-ghost:hover {
-      color: var(--text-primary);
-      background: rgba(255,255,255,0.04);
-    }
-
-    button:disabled {
-      opacity: 0.35;
-      cursor: not-allowed;
-      pointer-events: none;
-    }
-
-    .btn-row {
-      display: flex;
-      gap: 8px;
-    }
-
-    .btn-row button {
-      flex: 1;
-    }
-
-    /* Results */
-    .results {
-      margin-top: 16px;
-    }
-
-    /* App Item */
-    .app-item {
-      display: flex;
-      align-items: center;
-      padding: 14px;
-      background: var(--bg-elevated);
-      border: 1px solid var(--border);
-      border-radius: var(--radius);
-      margin-bottom: 10px;
-      gap: 12px;
-      flex-wrap: wrap;
-      transition: border-color var(--transition), background var(--transition);
-    }
-
-    .app-item:hover {
-      border-color: var(--border-focus);
-      background: var(--bg-card-hover);
-    }
-
-    .app-icon {
-      width: 44px;
-      height: 44px;
-      border-radius: 10px;
-      object-fit: cover;
-      background: var(--bg-primary);
-      flex-shrink: 0;
-      border: 1px solid var(--border);
-    }
-
-    .app-info {
-      flex: 1;
-      min-width: 150px;
-    }
-
-    .app-info h3 {
-      font-size: 14px;
-      font-weight: 600;
-      margin-bottom: 3px;
-      color: var(--text-primary);
-    }
-
-    .app-info .pkg {
-      font-size: 12px;
-      font-family: var(--font-mono);
-      color: var(--text-tertiary);
-      word-break: break-all;
-    }
-
-    .app-actions {
-      display: flex;
-      gap: 6px;
-    }
-
-    /* Messages */
-    .msg {
-      padding: 12px 14px;
-      border-radius: var(--radius);
-      margin-bottom: 10px;
-      font-size: 13px;
-      line-height: 1.5;
-      border: 1px solid;
-    }
-
-    .msg.err {
-      background: var(--error-bg);
-      color: var(--error);
-      border-color: rgba(248, 113, 113, 0.15);
-    }
-
-    .msg.ok {
-      background: var(--success-bg);
-      color: var(--success);
-      border-color: rgba(74, 222, 128, 0.15);
-    }
-
-    .msg.info {
-      background: var(--info-bg);
-      color: var(--info);
-      border-color: rgba(96, 165, 250, 0.12);
-    }
-
-    .msg strong {
-      color: var(--text-primary);
-      font-weight: 600;
-    }
-
-    .msg small {
-      opacity: 0.7;
-    }
-
-    /* Loading */
-    .loading {
-      text-align: center;
-      padding: 20px;
-      color: var(--text-secondary);
-      font-size: 13px;
-    }
-
-    .spinner {
-      display: inline-block;
-      width: 16px;
-      height: 16px;
-      border: 2px solid var(--border);
-      border-top-color: var(--accent);
-      border-radius: 50%;
-      animation: spin 0.8s linear infinite;
-      margin-right: 8px;
-      vertical-align: middle;
-    }
-
-    @keyframes spin { to { transform: rotate(360deg); } }
-
-    /* Progress Bar */
-    .progress {
-      width: 100%;
-      height: 4px;
-      background: var(--bg-primary);
-      border-radius: 2px;
-      margin-top: 10px;
-      overflow: hidden;
-    }
-
-    .progress-bar {
-      height: 100%;
-      background: linear-gradient(90deg, var(--accent-dim), var(--accent));
-      border-radius: 2px;
-      transition: width 0.3s ease;
-      position: relative;
-    }
-
-    .progress-bar::after {
-      content: '';
-      position: absolute;
-      inset: 0;
-      background: linear-gradient(90deg, transparent, rgba(255,255,255,0.2), transparent);
-      animation: shimmer 1.5s infinite;
-    }
-
-    @keyframes shimmer {
-      0% { transform: translateX(-100%); }
-      100% { transform: translateX(100%); }
-    }
-
-    /* Indeterminate progress */
-    .progress-indeterminate .progress-bar {
-      width: 40% !important;
-      animation: indeterminate 1.2s ease-in-out infinite;
-    }
-
-    @keyframes indeterminate {
-      0% { transform: translateX(-100%); }
-      100% { transform: translateX(350%); }
-    }
-
-    a { text-decoration: none; }
-    a button { width: 100%; }
-
-    /* Divider */
-    .divider {
-      height: 1px;
-      background: var(--border);
-      margin: 8px 0 20px;
-    }
-
-    /* Fade-in animation */
-    .fade-in {
-      animation: fadeIn 0.3s ease;
-    }
-
-    @keyframes fadeIn {
-      from { opacity: 0; transform: translateY(6px); }
-      to { opacity: 1; transform: translateY(0); }
-    }
-
-    /* Footer */
-    .footer {
-      text-align: center;
-      margin-top: 32px;
-      padding-top: 20px;
-      border-top: 1px solid var(--border);
-    }
-
-    .footer p {
-      font-size: 11px;
-      font-family: var(--font-mono);
-      color: var(--text-tertiary);
-      letter-spacing: 0.02em;
-    }
-
-    .footer-link {
-      display: inline-flex;
-      align-items: center;
-      gap: 6px;
-      margin-top: 10px;
-      font-family: var(--font-mono);
-      font-size: 12px;
-      font-weight: 500;
-      color: var(--accent);
-      text-decoration: none;
-      padding: 6px 14px;
-      border-radius: 20px;
-      border: 1px solid var(--accent-glow);
-      transition: all var(--transition);
-    }
-
-    .footer-link:hover {
-      color: var(--text-primary);
-      border-color: var(--accent);
-      background: var(--accent-glow);
-    }
-
-    .footer-link svg {
-      opacity: 0.7;
-      transition: opacity var(--transition);
-    }
-
-    .footer-link:hover svg {
-      opacity: 1;
-    }
-
-    /* ADB Card */
-    .adb-card {
-      border: 1px dashed var(--border);
-      background: var(--bg-card);
-    }
-
-    .adb-card.connected {
-      border-color: var(--success);
-      border-style: solid;
-      background: var(--success-bg);
-    }
-
-    .adb-status {
-      display: flex;
-      align-items: center;
-      gap: 10px;
-      min-height: 36px;
-    }
-
-    .adb-dot {
-      width: 8px;
-      height: 8px;
-      border-radius: 50%;
-      background: var(--success);
-      box-shadow: 0 0 6px var(--success);
-      flex-shrink: 0;
-    }
-
-    .adb-device-name {
-      font-family: var(--font-mono);
-      font-size: 12px;
-      color: var(--text-primary);
-      flex: 1;
-    }
-
-    .adb-device-name small {
-      display: block;
-      color: var(--text-secondary);
-      font-size: 11px;
-      margin-top: 2px;
-    }
-
-    .adb-unsupported {
-      font-size: 12px;
-      color: var(--text-tertiary);
-      font-family: var(--font-mono);
-    }
-
-    .btn-install {
-      background: #16a34a;
-      color: #fff;
-      border: none;
-      padding: 8px 16px;
-      border-radius: var(--radius);
-      font-family: var(--font-sans);
-      font-size: 13px;
-      font-weight: 600;
-      cursor: pointer;
-      display: inline-flex;
-      align-items: center;
-      gap: 6px;
-      transition: background var(--transition);
-    }
-
-    .btn-install:hover { background: #15803d; }
-    .btn-install:disabled { opacity: 0.5; cursor: not-allowed; }
-
-    .merge-tooltip {
-      position: absolute;
-      bottom: calc(100% + 8px);
-      left: 50%;
-      transform: translateX(-50%);
-      background: var(--bg-elevated);
-      border: 1px solid var(--border);
-      color: var(--text-secondary);
-      font-size: 11px;
-      padding: 8px 12px;
-      border-radius: var(--radius);
-      white-space: nowrap;
-      pointer-events: none;
-      opacity: 0;
-      transition: opacity var(--transition);
-      z-index: 10;
-    }
-
-    .merge-tooltip.show { opacity: 1; }
-
-    .checkbox-wrap { position: relative; }
-
-    /* Stat Counter */
-    .stat-counter {
-      margin-top: 12px;
-      font-family: var(--font-mono);
-      font-size: 12px;
-      color: var(--text-tertiary);
-      letter-spacing: 0.02em;
-    }
-
-    .stat-counter span {
-      color: var(--accent);
-      font-weight: 600;
-    }
-
-    /* Log Panel */
-    .log-panel {
-      background: var(--bg-card);
-      border: 1px solid var(--border);
-      border-radius: var(--radius-lg);
-      margin-bottom: 16px;
-      overflow: hidden;
-      transition: border-color var(--transition);
-    }
-
-    .log-panel:hover {
-      border-color: var(--border-focus);
-    }
-
-    .log-header {
-      display: flex;
-      align-items: center;
-      justify-content: space-between;
-      padding: 12px 16px;
-      cursor: pointer;
-      user-select: none;
-      transition: background var(--transition);
-    }
-
-    .log-header:hover {
-      background: rgba(255,255,255,0.02);
-    }
-
-    .log-header-left {
-      display: flex;
-      align-items: center;
-      gap: 8px;
-    }
-
-    .log-header .card-label {
-      margin-bottom: 0;
-    }
-
-    .log-badge {
-      font-family: var(--font-mono);
-      font-size: 10px;
-      font-weight: 600;
-      padding: 2px 7px;
-      border-radius: 10px;
-      background: var(--accent-glow);
-      color: var(--accent-dim);
-      min-width: 20px;
-      text-align: center;
-      transition: all var(--transition);
-    }
-
-    .log-badge.empty {
-      background: rgba(255,255,255,0.04);
-      color: var(--text-tertiary);
-    }
-
-    .log-toggle {
-      width: 20px;
-      height: 20px;
-      color: var(--text-tertiary);
-      transition: transform 0.2s ease;
-    }
-
-    .log-panel.open .log-toggle {
-      transform: rotate(180deg);
-    }
-
-    .log-body {
-      max-height: 0;
-      overflow: hidden;
-      transition: max-height 0.3s ease;
-    }
-
-    .log-panel.open .log-body {
-      max-height: 400px;
-    }
-
-    .log-scroll {
-      max-height: 300px;
-      overflow-y: auto;
-      padding: 0 16px 12px;
-      scroll-behavior: smooth;
-    }
-
-    .log-scroll::-webkit-scrollbar {
-      width: 4px;
-    }
-
-    .log-scroll::-webkit-scrollbar-track {
-      background: transparent;
-    }
-
-    .log-scroll::-webkit-scrollbar-thumb {
-      background: var(--border-focus);
-      border-radius: 2px;
-    }
-
-    .log-entry {
-      display: flex;
-      gap: 8px;
-      padding: 4px 0;
-      font-family: var(--font-mono);
-      font-size: 11px;
-      line-height: 1.5;
-      animation: logSlide 0.2s ease;
-      border-bottom: 1px solid rgba(255,255,255,0.02);
-    }
-
-    .log-entry:last-child {
-      border-bottom: none;
-    }
-
-    @keyframes logSlide {
-      from { opacity: 0; transform: translateY(-4px); }
-      to { opacity: 1; transform: translateY(0); }
-    }
-
-    .log-time {
-      color: var(--text-tertiary);
-      flex-shrink: 0;
-      opacity: 0.6;
-    }
-
-    .log-icon {
-      flex-shrink: 0;
-      width: 14px;
-      text-align: center;
-    }
-
-    .log-icon.info { color: var(--info); }
-    .log-icon.ok { color: var(--success); }
-    .log-icon.err { color: var(--error); }
-    .log-icon.warn { color: #fbbf24; }
-    .log-icon.dl { color: var(--accent); }
-
-    .log-msg {
-      color: var(--text-secondary);
-      word-break: break-word;
-    }
-
-    .log-empty {
-      padding: 16px 0;
-      text-align: center;
-      color: var(--text-tertiary);
-      font-family: var(--font-mono);
-      font-size: 11px;
-    }
-
-    .log-actions {
-      display: flex;
-      justify-content: flex-end;
-      padding: 0 16px 8px;
-    }
-
-    .log-clear-btn {
-      font-family: var(--font-mono) !important;
-      font-size: 10px !important;
-      padding: 4px 8px !important;
-      color: var(--text-tertiary) !important;
-      border: none !important;
-    }
-
-    /* Pulsing dot for active operations */
-    .log-active-dot {
-      width: 6px;
-      height: 6px;
-      border-radius: 50%;
-      background: var(--accent);
-      animation: pulse 1.5s ease-in-out infinite;
-    }
-
-    @keyframes pulse {
-      0%, 100% { opacity: 0.4; }
-      50% { opacity: 1; }
-    }
-
-    /* Backup/Restore */
-    .backup-card { display: none; }
-    .backup-card.visible { display: block; }
-
-    .backup-list {
-      max-height: 300px;
-      overflow-y: auto;
-      margin-top: 12px;
-      border: 1px solid var(--border);
-      border-radius: var(--radius);
-    }
-
-    .backup-item {
-      display: flex;
-      align-items: center;
-      gap: 10px;
-      padding: 8px 12px;
-      border-bottom: 1px solid var(--border);
-      font-size: 12px;
-      font-family: var(--font-mono);
-    }
-
-    .backup-item:last-child { border-bottom: none; }
-
-    .backup-item input[type="checkbox"] {
-      accent-color: var(--accent);
-      flex-shrink: 0;
-    }
-
-    .backup-item .pkg-name {
-      flex: 1;
-      color: var(--text-primary);
-      overflow: hidden;
-      text-overflow: ellipsis;
-      white-space: nowrap;
-    }
-
-    .backup-item .pkg-status {
-      font-size: 11px;
-      flex-shrink: 0;
-    }
-
-    .backup-item .pkg-status.available { color: var(--success); }
-    .backup-item .pkg-status.unavailable { color: var(--text-tertiary); }
-    .backup-item .pkg-status.checking { color: var(--info); }
-
-    .backup-summary {
-      font-size: 12px;
-      color: var(--text-secondary);
-      margin-top: 8px;
-      font-family: var(--font-mono);
-    }
-
-    .backup-actions {
-      display: flex;
-      gap: 8px;
-      margin-top: 16px;
-      padding-top: 4px;
-      flex-wrap: wrap;
-    }
-
-    #backup-result {
-      margin-top: 12px;
-    }
-
-    #backup-result .msg {
-      margin-bottom: 12px;
-    }
-
-    .file-input-wrap {
-      position: relative;
-      display: inline-flex;
-    }
-
-    .file-input-wrap input[type="file"] {
-      position: absolute;
-      inset: 0;
-      opacity: 0;
-      cursor: pointer;
-    }
-
-    /* Mobile adjustments */
-    @media (max-width: 480px) {
-      .wrapper { padding: 24px 16px 40px; }
-      .app-actions { width: 100%; }
-      .app-actions button { flex: 1; }
-      .btn-row button { padding: 10px 12px; font-size: 12px; }
-      .log-scroll { max-height: 200px; }
-    }
-  </style>
-</head>
-<body>
-  <div class="wrapper">
-    <header class="header">
-      <div class="header-badge">
-        <svg viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"><path d="M21 15v4a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2v-4"/><polyline points="7 10 12 15 17 10"/><line x1="12" y1="15" x2="12" y2="3"/></svg>
-        APK Downloader
-      </div>
-      <h1>G<span>Play</span></h1>
-      <p>Download APKs directly from Google Play Store</p>
-      <div class="stat-counter" id="stat-counter"></div>
-    </header>
-
-    <!-- ADB Device -->
-    <div class="card adb-card" id="adb-card">
-      <div class="card-label">
-        <svg viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"><rect x="4" y="2" width="16" height="20" rx="2"/><line x1="12" y1="18" x2="12.01" y2="18"/></svg>
-        Install to Device
-      </div>
-      <div class="adb-status" id="adb-status">
-        <button class="btn-secondary" id="adb-connect-btn" onclick="adbConnect()">
-          <svg width="14" height="14" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"><path d="M7 2v4M17 2v4M2 12h4M18 12h4M4.93 4.93l2.83 2.83M16.24 4.93l-2.83 2.83M12 8v4M12 16v.01"/><circle cx="12" cy="12" r="6"/></svg>
-          Connect Device
-        </button>
-      </div>
-    </div>
-
-    <!-- Download Card -->
-    <div class="card">
-      <div class="card-label">
-        <svg viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"><rect x="3" y="3" width="18" height="18" rx="2" ry="2"/><line x1="3" y1="9" x2="21" y2="9"/><line x1="9" y1="21" x2="9" y2="9"/></svg>
-        Direct Download
-      </div>
-      <input type="text" id="pkg-input" placeholder="com.example.app" spellcheck="false" autocomplete="off">
-      <div class="options-row">
-        <select id="arch-select">
-          <option value="arm64-v8a">ARM64 (modern)</option>
-          <option value="armeabi-v7a">ARMv7 (legacy)</option>
-        </select>
-        <label class="checkbox-wrap" id="merge-label">
-          <input type="checkbox" id="merge-apks" checked>
-          <span>Merge splits</span>
-        </label>
-      </div>
-      <div class="note">Merged APKs are re-signed with a debug key. Apps without splits keep their original signature.</div>
-      <div class="btn-row">
-        <button class="btn-secondary" onclick="getInfo()">
-          <svg width="14" height="14" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"><circle cx="12" cy="12" r="10"/><line x1="12" y1="16" x2="12" y2="12"/><line x1="12" y1="8" x2="12.01" y2="8"/></svg>
-          Info
-        </button>
-        <button class="btn-primary" onclick="download()">
-          <svg width="14" height="14" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"><path d="M21 15v4a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2v-4"/><polyline points="7 10 12 15 17 10"/><line x1="12" y1="15" x2="12" y2="3"/></svg>
-          Download
-        </button>
-      </div>
-      <div id="info-result" class="results"></div>
-    </div>
-
-    <!-- Log Panel -->
-    <div class="log-panel" id="log-panel">
-      <div class="log-header" onclick="toggleLog()">
-        <div class="log-header-left">
-          <div class="card-label">
-            <svg viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"><polyline points="4 17 10 11 4 5"/><line x1="12" y1="19" x2="20" y2="19"/></svg>
-            Activity Log
-          </div>
-          <span class="log-badge empty" id="log-badge">0</span>
-          <div class="log-active-dot" id="log-dot" style="display:none"></div>
-        </div>
-        <svg class="log-toggle" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"><polyline points="6 9 12 15 18 9"/></svg>
-      </div>
-      <div class="log-body">
-        <div class="log-actions">
-          <button class="log-clear-btn btn-ghost" onclick="event.stopPropagation();clearLog()">clear</button>
-        </div>
-        <div class="log-scroll" id="log-scroll">
-          <div class="log-empty" id="log-empty">No activity yet</div>
-        </div>
-      </div>
-    </div>
-
-    <!-- Search Card -->
-    <div class="card">
-      <div class="card-label">
-        <svg viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"><circle cx="11" cy="11" r="8"/><line x1="21" y1="21" x2="16.65" y2="16.65"/></svg>
-        Search
-      </div>
-      <input type="text" id="search-q" placeholder="Search the Play Store..." onkeypress="if(event.key==='Enter')search()">
-      <div class="btn-row">
-        <button class="btn-primary" onclick="search()">
-          <svg width="14" height="14" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"><circle cx="11" cy="11" r="8"/><line x1="21" y1="21" x2="16.65" y2="16.65"/></svg>
-          Search
-        </button>
-      </div>
-      <div id="search-results" class="results"></div>
-    </div>
-
-    <!-- Backup/Restore Card -->
-    <div class="card backup-card" id="backup-card">
-      <div class="card-label">
-        <svg viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"><path d="M19 21H5a2 2 0 0 1-2-2V5a2 2 0 0 1 2-2h11l5 5v11a2 2 0 0 1-2 2z"/><polyline points="17 21 17 13 7 13 7 21"/><polyline points="7 3 7 8 15 8"/></svg>
-        Backup &amp; Restore
-      </div>
-      <div class="backup-actions" id="backup-actions">
-        <button class="btn-primary" id="backup-btn" onclick="backupApps()" style="display:none">
-          <svg width="14" height="14" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"><path d="M21 15v4a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2v-4"/><polyline points="17 8 12 3 7 8"/><line x1="12" y1="3" x2="12" y2="15"/></svg>
-          Backup App List
-        </button>
-        <div class="file-input-wrap">
-          <button class="btn-secondary">
-            <svg width="14" height="14" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"><path d="M21 15v4a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2v-4"/><polyline points="7 10 12 15 17 10"/><line x1="12" y1="15" x2="12" y2="3"/></svg>
-            Import List
-          </button>
-          <input type="file" accept=".json" onchange="importBackup(event)">
-        </div>
-      </div>
-      <div id="backup-result"></div>
-    </div>
-
-    <footer class="footer">
-      <p>Uses Aurora Store token dispenser for anonymous auth</p>
-      <a href="https://github.com/alltechdev/gplay-apk-downloader" target="_blank" rel="noopener" class="footer-link">
-        <svg width="16" height="16" viewBox="0 0 24 24" fill="currentColor"><path d="M12 0C5.37 0 0 5.37 0 12c0 5.31 3.435 9.795 8.205 11.385.6.105.825-.255.825-.57 0-.285-.015-1.23-.015-2.235-3.015.555-3.795-.735-4.035-1.41-.135-.345-.72-1.41-1.23-1.695-.42-.225-1.02-.78-.015-.795.945-.015 1.62.87 1.845 1.23 1.08 1.815 2.805 1.305 3.495.99.105-.78.42-1.305.765-1.605-2.67-.3-5.46-1.335-5.46-5.925 0-1.305.465-2.385 1.23-3.225-.12-.3-.54-1.53.12-3.18 0 0 1.005-.315 3.3 1.23.96-.27 1.98-.405 3-.405s2.04.135 3 .405c2.295-1.56 3.3-1.23 3.3-1.23.66 1.65.24 2.88.12 3.18.765.84 1.23 1.905 1.23 3.225 0 4.605-2.805 5.625-5.475 5.925.435.375.81 1.095.81 2.22 0 1.605-.015 2.895-.015 3.3 0 .315.225.69.825.57A12.02 12.02 0 0024 12c0-6.63-5.37-12-12-12z"/></svg>
-        alltechdev/gplay-apk-downloader
-      </a>
-    </footer>
-  </div>
-
-  <script type="module">
-    import { Adb, AdbDaemonTransport } from 'https://cdn.jsdelivr.net/npm/@yume-chan/adb@2.1.0/+esm';
-    import { AdbDaemonWebUsbDeviceManager } from 'https://cdn.jsdelivr.net/npm/@yume-chan/adb-daemon-webusb@2.1.0/+esm';
-    import AdbWebCredentialStore from 'https://cdn.jsdelivr.net/npm/@yume-chan/adb-credential-web@2.1.0/+esm';
-
-    class AdbManager {
-      constructor() {
-        this.manager = AdbDaemonWebUsbDeviceManager.BROWSER;
-        this.device = null;
-        this.transport = null;
-        this.adb = null;
-        this.credentialStore = new AdbWebCredentialStore('GPlayADB');
-      }
-
-      get supported() { return !!this.manager; }
-      get connected() { return !!this.adb; }
-
-      async connect() {
-        if (!this.manager) throw new Error('WebUSB not supported');
-
-        // Clean up any previous connection
-        await this.disconnect();
-
-        // Try cached/authorized devices first
-        const devices = await this.manager.getDevices();
-        let target = null;
-        const cached = localStorage.getItem('adbDevice');
-        if (cached) {
-          const info = JSON.parse(cached);
-          target = devices.find(d => d.serial === info.serial);
-        }
-        if (!target && devices.length > 0) target = devices[0];
-
-        // Try connecting to cached device
-        if (target) {
-          try {
-            return await this._connectDevice(target);
-          } catch (e) {
-            // Cached device failed, clear and show picker
-            localStorage.removeItem('adbDevice');
-            target = null;
-          }
-        }
-
-        // Show device picker
-        target = await this.manager.requestDevice();
-        if (!target) throw new Error('No device selected');
-        return await this._connectDevice(target);
-      }
-
-      async _connectDevice(target) {
-        localStorage.setItem('adbDevice', JSON.stringify({ serial: target.serial }));
-        const connection = await target.connect();
-        this.transport = await AdbDaemonTransport.authenticate({
-          serial: target.serial,
-          connection,
-          credentialStore: this.credentialStore
-        });
-        this.adb = new Adb(this.transport);
-        this.device = target;
-        return await this.getDeviceInfo();
-      }
-
-      async disconnect() {
-        if (this.adb) { try { await this.adb.close(); } catch {} }
-        this.adb = null;
-        this.transport = null;
-        this.device = null;
-      }
-
-      async getDeviceInfo() {
-        const [model, android] = await Promise.all([
-          this.shell('getprop ro.product.model'),
-          this.shell('getprop ro.build.version.release')
-        ]);
-        return { model: model.trim(), android: android.trim() };
-      }
-
-      async shell(cmd) {
-        return await this.adb.subprocess.noneProtocol.spawnWaitText(cmd);
-      }
-
-      async pushFile(blob, remotePath) {
-        const sync = await this.adb.sync();
-        try {
-          await sync.write({ filename: remotePath, file: blob.stream(), permission: 0o644 });
-        } finally {
-          await sync.dispose();
-        }
-      }
-
-      async installSingle(blob, filename) {
-        const tmpPath = '/data/local/tmp/gplay_' + Date.now() + '.apk';
-        await this.pushFile(blob, tmpPath);
-        try {
-          const result = await this.shell('pm install -r "' + tmpPath + '"');
-          if (!result.includes('Success')) throw new Error(result.trim());
-        } finally {
-          await this.shell('rm -f "' + tmpPath + '"');
-        }
-      }
-
-      async installSplit(apks, onProgress) {
-        const totalSize = apks.reduce((s, a) => s + a.size, 0);
-        if (onProgress) onProgress('create', 'Creating install session...');
-        const createOut = await this.shell('pm install-create -S ' + totalSize);
-        const match = createOut.match(/\[(\d+)\]/);
-        if (!match) throw new Error('Failed to create session: ' + createOut.trim());
-        const sessionId = match[1];
-
-        try {
-          for (let i = 0; i < apks.length; i++) {
-            const apk = apks[i];
-            const tmpPath = '/data/local/tmp/gplay_' + Date.now() + '_' + i + '.apk';
-            if (onProgress) onProgress('push', 'Pushing ' + apk.name + ' (' + (i + 1) + '/' + apks.length + ')...');
-            await this.pushFile(apk.blob, tmpPath);
-
-            if (onProgress) onProgress('write', 'Writing ' + apk.name + ' to session...');
-            const label = i === 0 ? 'base.apk' : 'split_' + i + '.apk';
-            const writeOut = await this.shell('pm install-write -S ' + apk.size + ' ' + sessionId + ' ' + label + ' "' + tmpPath + '"');
-            if (!writeOut.includes('Success')) throw new Error('install-write failed for ' + label + ': ' + writeOut.trim());
-            await this.shell('rm -f "' + tmpPath + '"');
-          }
-
-          if (onProgress) onProgress('commit', 'Committing install...');
-          const commitOut = await this.shell('pm install-commit ' + sessionId);
-          if (!commitOut.includes('Success')) throw new Error('Install failed: ' + commitOut.trim());
-        } catch (e) {
-          try { await this.shell('pm install-abandon ' + sessionId); } catch {}
-          throw e;
-        }
-      }
-    }
-
-    window.adbManager = new AdbManager();
-  </script>
-
-  <script>
     const $ = s => document.querySelector(s);
 
     let downloadEventSource = null;
@@ -2075,6 +859,3 @@
       if (statusEl) statusEl.textContent = current + '/' + total;
       if (barEl) barEl.style.width = pct + '%';
     }
-  </script>
-</body>
-</html>
